# swagger.yaml
swagger: '2.0'
schemes:
  - http
host: 'localhost:3000'
basePath: /
info:
  version: '1.0'
  title: products-API_with_node_js
  description: Administa o catalógo e estoque de produtos de um e-commerce
  contact:
    name: Suporte
    email: vitor.zogbi_bols@compasso.com.br
consumes:
  - application/json
produces:
  - application/json
tags:
  - name: skuproduct
  - name: sku
  - name: productswithsku
paths:
  '/skuproducts/create':
    post:
      tags:
        - skuproduct
      summary: Cadastra produto
      description: Cadastra um novo produto que possua skus
      operationId: createProduct
      parameters:
        - in: body
          name: skuproducts
          required: true
          schema:
            $ref: '#/definitions/skuproducts'
      responses:
        '201':
          description: Produto cadastrado com sucesso
          schema:
            $ref: '#/definitions/skuproducts'
        '422':
          description: Erro com os dados do produto
        '500':
          description: Não foi possível cadastrar o produto
  '/skuproducts/list':
    get:
      tags:
        - skuproduct
      summary: Listagem de produtos
      description: >-
        Retorna uma listagem com todos os produtos e suas skus cadastrados na
        base de dados
      operationId: listProducts
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/skuproducts'
        '500':
          description: Não foi possível consultar os produtos
  '/skuproducts/listpaginated/{page}':
    get:
      tags:
        - skuproduct
      summary: Listagem de produtos paginada
      description: Retorna listagem dos produtos com paginação
      parameters:
        - in: path
          name: page
          type: string
          required: true
        - in: query
          name: skuproduct
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/skuproducts'
        '206':
          description: Zero produto nessa página
        '500':
          description: Falha ao listar os produtos
  '/skuproducts/findbyid/{id}':
    get:
      tags:
        - skuproduct
      summary: Encontra produto pela id
      description: Retorna os dados de um produto com base na id informada
      operationId: findProductById
      parameters:
        - in: path
          name: id
          type: string
          required: true
        - in: query
          name: skuproduct
          type: string
      responses:
        '200':
          description: Produto encontrado
          schema:
            $ref: '#/definitions/skuproducts'
        '404':
          description: Produto não encontrado
        '422':
          description: Erro com os dados do produto
        '500':
          description: Falha ao encontrar o produto
  '/skuproducts/update/{id}':
    put:
      tags:
        - skuproduct
      summary: Altera produto
      description: Altera algum(ns) valor(s) do produto
      operationId: updateProduct
      parameters:
        - in: path
          name: id
          type: string
          required: true
        - in: body
          name: skuproducts
          required: true
          schema:
            $ref: '#/definitions/skuproducts'
        - in: query
          name: skuproduct
          type: string
      responses:
        '201':
          description: Produto atualizado
          schema:
            $ref: '#/definitions/skuproducts'
        '422':
          description: Erro com os dados do produto
        '500':
          description: Não foi possível alterar o produto
  '/skuproducts/delete/{id}':
    delete:
      tags:
        - skuproduct
      summary: Exclui o produto
      description: Exclusão de um produto através do Id
      operationId: deleteProduct
      parameters:
        - in: path
          name: id
          type: string
          required: true
        - in: query
          name: skuproduct
          type: string
      responses:
        '204':
          description: Produto apagado com sucesso
        '404':
          description: Produto não encontrado
        '422':
          description: Erro com os dados do produto
        '500':
          description: Não foi possível apagar o produto desejado
  '/sku/create':
    post:
      tags:
        - sku
      summary: Cria uma sku
      description: Cria uma sku referente a um produto
      operationId: createSku
      parameters:
        - in: body
          name: skus
          required: true
          schema:
            $ref: '#/definitions/skus'
      responses:
        '201':
          description: ok
        '404':
          description: SKU não criada
        '422':
          description: Erro com os dados da sku
        '500':
          description: Não foi possível criar a sku
  '/sku/list':
    get:
      tags:
        - sku
      summary: Lista skus
      description: Lista todas as skus cadastradas no sistema
      operationId: listSku
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/skus'
        '500':
          description: Falha ao carregar as skus
  '/sku/listpaginated/{page}':
    get:
      tags:
        - sku
      summary: Listagem de skus paginada
      description: Retorna listagem das skus com paginação
      operationId: listPaginated
      parameters:
        - in: path
          name: page
          type: string
          required: true
        - in: query
          name: sku
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/skus'
        '206':
          description: Nenhuma sku encontrada
        '500':
          description: Falha ao listar os produtos
  '/sku/findbyid/{id}':
    get:
      tags:
        - sku
      summary: Encontra sku pela id
      description: Retorna os dados de uma sku com base na id informada
      operationId: findSkuById
      parameters:
        - in: path
          name: id
          type: string
          required: true
        - in: query
          name: sku
          type: string
      responses:
        '200':
          description: Sku encontrada
          schema:
            $ref: '#/definitions/skus'
        '404':
          description: Sku não encontrada
        '422':
          description: Erro com a id
        '500':
          description: Falha ao encontrar a sku
  '/sku/findbyproductid/{id}':
    get:
      tags:
        - sku
      summary: Encontra skus pela id de produto
      description: Retorna todas as skus de um produto baseada na id do produto
      operationId: findSkuByProductId
      parameters:
        - in: path
          name: id
          type: string
          required: true
        - in: query
          name: sku
          type: string
      responses:
        '200':
          description: Sku encontrada
          schema:
            $ref: '#/definitions/skus'
        '404':
          description: Sku não encontrada
        '422':
          description: Erro com a id
        '500':
          description: Falha ao encontrar a sku
  '/sku/update/{id}':
    put:
      tags:
        - sku
      summary: Altera uma sku
      description: Altera uma sku com base na id
      operationId: updateSku
      parameters:
        - in: path
          name: id
          type: string
          required: true
        - in: body
          name: sku
          required: true
          schema:
            $ref: '#/definitions/skus'
      responses:
        '201':
          description: Sku editada com sucesso
          schema:
            $ref: '#/definitions/skus'
        '404':
          description: Sku não encontrada
        '422':
          description: Erro com os dados da sku
        '500':
          description: Falha ao encontrar a sku
  '/sku/delete/{id}':
    delete:
      tags:
        - sku
      summary: Exclui a sku
      description: Exclusão de uma sku através da Id
      operationId: deleteSku
      parameters:
        - in: path
          name: id
          type: string
          required: true
        - in: query
          name: sku
          type: string
      responses:
        '204':
          description: Sku apagada com sucesso
        '404':
          description: Sku não encontrada
        '422':
          description: Erro com os dados da sku
        '500':
          description: Não foi possível apagar a sku desejada
  
  '/productswithsku/createproduct':
    post:
      tags:
        - productswithsku
      summary: Cria um produto
      description: Cria um novo produto
      operationId: createProductWithSku
      parameters:
        - in: body
          name: product
          required: true
          schema:
            $ref: '#/definitions/productwithsku'
      responses:
        '201':
          description: ok
        '404':
          description: Produto não criado
        '422':
          description: Erro com os dados do produto
        '500':
          description: Não foi possível criar o produto        
  
  '/productswithsku/createsku':
    post:
      tags:
        - productswithsku
      summary: Cria uma sku
      description: Cria uma sku referente a um produto
      operationId: createSkuInProduct
      parameters:
        - in: body
          name: sku
          required: true
          schema:
            $ref: '#/definitions/productwithsku'
      responses:
        '201':
          description: ok
        '404':
          description: SKU não criada
        '422':
          description: Erro com os dados da sku
        '500':
          description: Não foi possível criar a sku
          
  '/productswithsku/list':
    get:
      tags:
        - productswithsku
      summary: Lista os produtos
      description: Lista os produstos com suas skus
      operationId: listProductWithSku
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/productwithsku'
        '500':
          description: Falha ao carregar os produtos
  
  '/productswithsku/list/{page}':
    get:
      tags:
        - productswithsku
      summary: Lista os produtos com paginação
      description: Lista os produtos com suas skus em formato paginado
      operationId: listProductsWithSkuPaginated
      parameters:
        - in: path
          name: page
          type: string
          required: true
      responses:
        '200':
          description: Produto(s) encontrado(s)
          schema:
            $ref: '#/definitions/productwithsku'
        '206':
          description: Nenhum produto encontrado
        '500':
          description: Falha ao carregar os produtos
  
  '/productswithsku/findbyid/{id}':
    get:
      tags:
        - productswithsku
      summary: Encontra um produto através de sua Id
      description: Retorna um produto com suas skus
      operationId: findById
      parameters:
        - in: path
          name: id
          type: string
          required: true
        - in: query
          name: productwithsku
          type: string
      responses:
        '200':
          description: Produto encontrado
          schema:
            $ref: '#/definitions/productwithsku'
        '404':
          description: Produto não encontrado
        '422':
          description: Erro com a id do produto
        '500':
          description: Falha ao encontrar o produto  
          
  '/productswithsku/update/{id}':
    put:
      tags:
        - productswithsku
      summary: Atualiza um produto
      description: Atualiza os dados base de um produto com base na sua id
      operationId: update
      parameters:
        - in: path
          name: id
          type: string
          required: true
        - in: body
          name: productswithsku
          required: true
          schema:
            $ref: '#/definitions/productwithsku'
        - in: query
          name: productwithsku
          type: string
      responses:
        '201':
          description: Produto atualizado
          schema:
            $ref: '#/definitions/productwithsku'
        '422':
          description: Erro com os dados do produto
        '500':
          description: Não foi possível alterar o produto
  '/productswithsku/updatesku/{id}':
    put:
      tags:
        - productswithsku
      summary: Atualiza uma sku
      description: Atualiza os dados de uma sku com base na sua id
      operationId: updateSkuInProducts
      parameters:
        - in: path
          name: id
          type: string
          required: true
        - in: body
          name: productswithsku
          required: true
          schema:
            $ref: '#/definitions/productwithsku'
        - in: query
          name: productwithsku
          type: string
      responses:
        '201':
          description: Sku atualizada
          schema:
            $ref: '#/definitions/productwithsku'
        '422':
          description: Erro com os dados da sku
        '500':
          description: Não foi possível alterar a sku
  '/productswithsku/delete/{id}':
    delete:
      tags:
        - productswithsku
      summary: Exclui o produto
      description: Exclusão de um produto através do Id
      operationId: delete
      parameters:
        - in: path
          name: id
          type: string
          required: true
        - in: query
          name: skuproduct
          type: string
      responses:
        '204':
          description: Produto apagado com sucesso
        '404':
          description: Produto não encontrado
        '422':
          description: Erro com os dados do produto
        '500':
          description: Não foi possível apagar o produto desejado
  '/productswithsku/deletesku/{id}':
    delete:
      tags:
        - productswithsku
      summary: Exclui a sku
      description: Exclusão de uma sku de um produto através do Id da sku
      operationId: deleteSkuInProduct
      parameters:
        - in: path
          name: id
          type: string
          required: true
        - in: query
          name: productwithsku
          type: string
      responses:
        '204':
          description: Produto apagado com sucesso
        '404':
          description: Produto não encontrado
        '422':
          description: Erro com os dados do produto
        '500':
          description: Não foi possível apagar o produto desejado
definitions:
  skuproducts:
    type: object
    properties:
      _id:
        type: string
        example: 5d826e5d57eab10780aecd1d
      active:
        type: boolean
      name:
        type: string
        format: utf-8
        example: Camisa
      description:
        type: string
        format: utf-8
        example: Item para vestir o torso
  skus:
    type: object
    properties:
      _id:
        type: string
        example: 5d826e5d57eab10780aecd1e
      productId:
        $ref: '#/definitions/skuproducts'
      color:
        type: string
        format: utf-8
        example: verde
      size:
        type: string
        format: utf-8
        example: GG
      price:
        type: number
        format: double
        example: 49.9
      stockLevel:
        type: number
        format: integer
        example: 10
      active:
        type: boolean
  productwithsku:
    type: object
    properties:
      _id:
        type: string
        example: 5d9276e9972aa819ecfc8684
      name:
        type: string
        format: utf-8
        example: Camisa
      description:
        type: string
        example: "Item de vestúario"
      active:
        type: boolean
        default: false
      sku:
        type: array
        items:
          type: object
          properties:
            color:
              type: string
              format: utf-8
              example: rose
            size:
              type: string
              format: utf-8
              example: rose
            price:
              type: number
              format: double
              example: 99.99
            stockLevel:
              type: number
              format: integer
              example: 10
            active:
              type: boolean